+++ gh-pages/js/examples-renderer.js
@@ -52368,7 +52368,7 @@ function singleInputTask(taskName) {
     return globalTask.length === 1;
   }
 
-  var singleInputTasks = ['return', 't'];
+  var singleInputTasks = ['alert', 'return', 't'];
 
   return singleInputTasks.indexOf(taskName) > -1;
 }
@@ -52890,11 +52890,16 @@ exports.head = function () {
   return document.head
 }
 
-exports.innerHTML = function (node, content) {
-  node.innerHTML = content
+// TODO more tags
+var tags = [
+  'a', 'div', 'link', 'p', 'script'
+]
 
-  return node
-}
+tags.forEach(function (x) {
+  exports[x] = function () {
+    return document.createElement(x)
+  }
+})
 
 exports.window = function () {
   return window